+++ src/Formatter/Coder.php
@@ -18,11 +18,6 @@ final class Coder implements Formatter
     private $codeMap;
 
     /**
-     * @var CharTree[] Memoization of formatted character trees.
-     */
-    private $memoization = [];
-
-    /**
      * @param CodeMap $codeMap Code map for coding character trees.
      */
     public function __construct(CodeMap $codeMap)
@@ -36,23 +31,7 @@ final class Coder implements Formatter
      */
     protected function applyCurrent(CharTree $charTree): CharTree
     {
-        return $this->applyInternal($charTree);
-    }
-
-    /**
-     * @param CharTree $charTree Character tree to format.
-     * @return CharTree Coded variant of the character tree.
-     */
-    private function applyInternal(CharTree $charTree): CharTree
-    {
-        // When PHP 7.1 is no longer supported, change to using spl_object_id.
-        $hash = spl_object_hash($charTree);
-
-        if (!isset($this->memoization[$hash])) {
-            $this->memoization[$hash] = $this->format($charTree);
-        }
-
-        return $this->memoization[$hash];
+        return $this->format($charTree);
     }
 
     /**
@@ -68,7 +47,7 @@ final class Coder implements Formatter
         $branches = [];
 
         foreach ($this->codeMap->getMap($charTree) as $char => $codedChars) {
-            $branch = $this->applyInternal($charTree->getBranchesAfterRoot((string)$char));
+            $branch = $this->applyCurrent($charTree->getBranchesAfterRoot((string)$char));
 
             foreach ($codedChars as $codedChar) {
                 $branches[] = CharTree::fromString($codedChar, [$branch]);